refactor(combineLatest): one file, moved patch file
- combined impl and support into one file
- moved patch file to add/observable/combineLatest

BREAKING CHANGE: add/operator/combineLatest-static is now add/observable/combineLatest
