+++ src/Rx.KitchenSink.ts
@@ -23,7 +23,7 @@ export interface KitchenSinkOperators<T> extends CoreOperators<T> {
 
 // statics
 /* tslint:disable:no-use-before-declare */
-import './add/operator/combineLatest-static';
+import './add/observable/combineLatest';
 import './add/operator/concat-static';
 import './add/operator/merge-static';
 import './add/observable/race';
