+++ src/add/observable/combineLatest.ts
@@ -1,0 +1,6 @@
+import {Observable} from '../../Observable';
+import {combineLatestStatic} from '../../operator/combineLatest';
+
+Observable.combineLatest = combineLatestStatic;
+
+export var _void: void;
