--- src/add/operator/combineLatest-static.ts
@@ -1,7 +1,0 @@
-
-import {Observable} from '../../Observable';
-import {combineLatest} from '../../operator/combineLatest-static';
-
-Observable.combineLatest = combineLatest;
-
-export var _void: void;
